move telemetry service inside monitoring module
