+++ packages/twenty-server/src/modules/messaging/message-import-manager/messaging-import-manager.module.ts
@@ -33,6 +33,7 @@ import { MessagingMessagesImportService } from 'src/modules/messaging/message-im
 import { MessagingPartialMessageListFetchService } from 'src/modules/messaging/message-import-manager/services/messaging-partial-message-list-fetch.service';
 import { MessagingSaveMessagesAndEnqueueContactCreationService } from 'src/modules/messaging/message-import-manager/services/messaging-save-messages-and-enqueue-contact-creation.service';
 import { MessageParticipantManagerModule } from 'src/modules/messaging/message-participant-manager/message-participant-manager.module';
+import { MessagingMonitoringModule } from 'src/modules/messaging/monitoring/messaging-monitoring.module';
 
 @Module({
   imports: [
@@ -47,6 +48,7 @@ import { MessageParticipantManagerModule } from 'src/modules/messaging/message-p
     EmailAliasManagerModule,
     FeatureFlagModule,
     MessageParticipantManagerModule,
+    MessagingMonitoringModule,
   ],
   providers: [
     MessagingMessageListFetchCronCommand,
