+++ packages/twenty-server/src/modules/messaging/monitoring/messaging-monitoring.module.ts
@@ -1,15 +1,18 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
+import { AnalyticsModule } from 'src/engine/core-modules/analytics/analytics.module';
+import { BillingModule } from 'src/engine/core-modules/billing/billing.module';
 import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 import { DataSourceEntity } from 'src/engine/metadata-modules/data-source/data-source.entity';
 import { MessagingCommonModule } from 'src/modules/messaging/common/messaging-common.module';
 import { MessagingMessageChannelSyncStatusMonitoringCronCommand } from 'src/modules/messaging/monitoring/crons/commands/messaging-message-channel-sync-status-monitoring.cron.command';
 import { MessagingMessageChannelSyncStatusMonitoringCronJob } from 'src/modules/messaging/monitoring/crons/jobs/messaging-message-channel-sync-status-monitoring.cron';
-import { BillingModule } from 'src/engine/core-modules/billing/billing.module';
+import { MessagingTelemetryService } from 'src/modules/messaging/monitoring/services/messaging-telemetry.service';
 
 @Module({
   imports: [
+    AnalyticsModule,
     MessagingCommonModule,
     BillingModule,
     TypeOrmModule.forFeature([Workspace], 'core'),
@@ -18,7 +21,8 @@ import { BillingModule } from 'src/engine/core-modules/billing/billing.module';
   providers: [
     MessagingMessageChannelSyncStatusMonitoringCronCommand,
     MessagingMessageChannelSyncStatusMonitoringCronJob,
+    MessagingTelemetryService,
   ],
-  exports: [],
+  exports: [MessagingTelemetryService],
 })
 export class MessagingMonitoringModule {}
